Add better error logging to live-init
We do some backflips to capture stderr output in the bulk of the code.
We redirect it to the file /bootstrap.errors.  We launch "tail | sed |
tee" in the bg to color these error lines and send them to stdout as
well as put them in the log file.

Actually there is a "while read line; do ... done" in the bg job as well
so sed and tee are called once for each line of output.  It was tricky
to get this set up.  The simplier, more obvious solutions didn't work.
But there is probably a much simpler way that does work.