Correct logic for when we make the homefs file
Always make homefs if it is required (homefs!)

Also make if it is wanted (homefs) and rootfs is required (rootfs!) and
no rootfs or rootfs.new file exists (before we try to make it).

The idea is to offer to make homefs just once if rootfs!,homefs but
don't offer after the first time otherwise it is a nag.