Fix obvious bug in umounting the extra fbsplash sys directory
fbcondecor_helper mounted a sys directory under /lib and this caused a
flood of error messages when I prepared to switch_root.  I fixed the
problem in fbcondecor_helper but wanted to leave this fix in live-init
but it has not been exercised recently and it had the io redirect in the
wrong place.